Fixed the remove button bug and add new components

+++ modulo2/entrega-apis-labenusers/apis-labenusers/src/App.js
@@ -1,28 +1,17 @@
 import React from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
+import LoginForm from './components/LoginForm/LoginForm.js';
+import UsersList from './components/UsersList/UsersList.js';
 
 const Page = styled.div`
     text-align: center;
-
-    button{
-        margin: 20px;
-    }
-`;
-const Form = styled.form`
-    margin: 10px auto;
     display: flex;
     flex-direction: column;
     align-items: center;
-    height: 150px;
-    padding: 10px;
-    box-sizing: border-box;
-    border: 1px black solid;
-    width: fit-content;
-    justify-content: space-around; 
 
-    input{
-        width: 180px;
+    button{
+        margin: 20px;
     }
 `;
 
@@ -64,37 +53,54 @@ class App extends React.Component {
             alert('Tente outra vez!');
         });
 
+        this.setState({inputNameValue: '', inputMailValue: ''});
     };
     changePage = () => {
         this.state.currentPage === true ? this.setState({currentPage: false}) : this.setState({currentPage: true});
-
-        console.log(this.state.usersList);
     };
-    componentDidUpdate = () => {
+    // getUsers = () => {
+    //     if (!this.state.currentPage) {
+    //         const url = 'https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users';
+            
+    //         const header = {
+    //             headers: {
+    //                 Authorization: "leo-christen-moreira"
+    //             }
+    //         };
+
+    //         axios.get(url, header)
+    //         .then((res) => {
+    //             return this.setState({usersList: res.data});
+    //         })
+    //         .catch((err) => {
+    //             console.log(err.res);
+    //         });
+    //     }
+    // };
+    getUsers = async () => {
         if (!this.state.currentPage) {
             const url = 'https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users';
-            
+                    
             const header = {
                 headers: {
                     Authorization: "leo-christen-moreira"
                 }
             };
 
-            axios.get(url, header)
-            .then((res) => {
-                return this.setState({usersList: res.data});
-            })
-            .catch((err) => {
-                console.log(err);
-            });
-        }
+            try {
+                const res = await axios.get(url, header);
+
+                this.setState({usersList: res.data})
+            }
+            catch(err) {
+                console.log(err.res)
+            };
+        }    
     };
-    removeUser = (evt) => {
-
-        this.setState({userId: evt.target.id});
-        console.log(typeof evt.target.id);
-
-        const id = this.state.userId;
+    componentDidUpdate = () => {
+        this.getUsers()
+    };
+    removeUser = (id) => {
 
         const url = `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`;
 
@@ -104,64 +110,67 @@ class App extends React.Component {
             }
         };
 
-        axios.delete(url, header)
-        .then((res) => {
-            alert('Usuário deletado com sucesso!');
-        })
-        .catch((err) => {
-            console.log(this.state.userId)
-            console.log(err.message);
-            alert('Opa, algo deu errado tente outra vez');
-        });
+        const answer = window.confirm('Você gostaria mesmo de deletar esse usuário?');
+
+        if (answer) {
+            axios.delete(url, header)
+            .then((res) => {
+                alert('Usuário deletado com sucesso!');
+            })
+            .catch((err) => {
+                console.log(this.state.userId)
+                console.log(err.message);
+                alert('Opa, algo deu errado tente outra vez');
+            });
+        }
+        else {
+            alert('O usuário não foi deletado.');
+        }
 
     };
 
     render() {
 
-        const montaCard = () => {
-            const list = this.state.usersList;
-
-            const user = list.map((user) => {
-                return (
-                    <li key={user.id}>{user.name} <button id={user.id} onClick={this.removeUser}>x</button></li>
-                );
-            });
-
-            return user;
-        }
-
         const pagesShift = () => {
                 if (this.state.currentPage) {
                     return (
-                        <Form>
-                            <div>
-                                <label htmlFor="name">Nome: </label>
-                                <input type="text" name='name'placeholder='digite seu nome aqui' value={this.state.inputNameValue} onChange={this.changeNameValue}/>
-                            </div>    
-        
-                            <div>
-                                <label htmlFor="mail">Email: </label>
-                                <input type="text" name='mail' placeholder='digite aqui seu email' value={this.state.inputMailValue} onChange={this.changeMailValue}/>
-                            </div>    
-        
-                            <button type='button' onClick={this.addUsr}>Adicionar usuário</button>
-                        </Form>
+                        <LoginForm 
+                            inputNameValue={this.state.inputNameValue}
+                            inputMailValue={this.state.inputMailValue}
+                            changeNameValue={this.changeNameValue}
+                            changeMailValue={this.changeMailValue}
+                            addUsr={this.addUsr}
+                        />
                     );
                 }
                 else {
                     return (
-                        <ul>
-                           {montaCard()} 
-                        </ul>
+                        <UsersList 
+                            usersList={this.state.usersList}
+                            removeUser={this.removeUser}
+                        />
                     )
                 }
+        };
+
+        const changeBttn = () => {
+            if (this.state.currentPage) {
+                return(
+                    <button type='button' onClick={this.changePage}>Ver usuários</button>
+                );
+            }
+            else {
+                return(
+                    <button type='button' onClick={this.changePage}>Adicionar novo usuário</button>
+                );
+            }
         };
 
         return (
             <Page>
                 <h1>Labenusers</h1>
                 {pagesShift()}
-                <button type='button' onClick={this.changePage}>Ver usuários</button>
+                {changeBttn()}
             </Page>
         );
     };
